--- nlprule/configs/en/tokenizer.json
@@ -1,9 +1,0 @@
-{
-    "allow_errors": false,
-    "ignore_ids": [
-        "DISAMBIGUATION/BEST_JJS/0"
-    ],
-    "extra_join_regexes": [
-        "(https?:\\/\\/(?:www\\.|(?!www))[a-zA-Z0-9][a-zA-Z0-9-]+[a-zA-Z0-9]\\.[^\\s]{2,}|www\\.[a-zA-Z0-9][a-zA-Z0-9-]+[a-zA-Z0-9]\\.[^\\s]{2,}|https?:\\/\\/(?:www\\.|(?!www))[a-zA-Z0-9]+\\.[^\\s]{2,}|www\\.[a-zA-Z0-9]+\\.[^\\s]{2,})"
-    ]
-}
