+++ nlprule/src/bin/test.rs
@@ -1,43 +1,45 @@
-use clap::Clap;
-use nlprule::{rules::Rules, tokenizer::Tokenizer};
+// use clap::Clap;
+// use nlprule::{rules::Rules, tokenizer::Tokenizer};
 
-#[derive(Clap)]
-#[clap(
-    version = "1.0",
-    author = "Benjamin Minixhofer <bminixhofer@gmail.com>"
-)]
-struct Opts {
-    #[clap(long, short)]
-    tokenizer: String,
-    #[clap(long, short)]
-    rules: String,
-    #[clap(long, short)]
-    ids: Vec<String>,
-}
+// #[derive(Clap)]
+// #[clap(
+//     version = "1.0",
+//     author = "Benjamin Minixhofer <bminixhofer@gmail.com>"
+// )]
+// struct Opts {
+//     #[clap(long, short)]
+//     tokenizer: String,
+//     #[clap(long, short)]
+//     rules: String,
+//     #[clap(long, short)]
+//     ids: Vec<String>,
+// }
 
-fn main() {
-    env_logger::init();
-    let opts = Opts::parse();
+// fn main() {
+//     env_logger::init();
+//     let opts = Opts::parse();
 
-    let tokenizer = Tokenizer::new(opts.tokenizer).unwrap();
-    let rules_container = Rules::new(opts.rules).unwrap();
-    let rules = rules_container.rules();
+//     let tokenizer = Tokenizer::new(opts.tokenizer).unwrap();
+//     let rules_container = Rules::new(opts.rules).unwrap();
+//     let rules = rules_container.rules();
 
-    println!("Runnable rules: {}", rules.len());
+//     println!("Runnable rules: {}", rules.len());
 
-    let mut passes = 0;
-    for rule in rules {
-        if opts.ids.is_empty() || opts.ids.contains(&rule.id().to_string()) {
-            if let Ok(true) = rule.test(&tokenizer) {
-                passes += 1;
-            }
-        }
-    }
+//     let mut passes = 0;
+//     for rule in rules {
+//         if opts.ids.is_empty() || opts.ids.contains(&rule.id().to_string()) {
+//             if let Ok(true) = rule.test(&tokenizer) {
+//                 passes += 1;
+//             }
+//         }
+//     }
 
-    println!("Rules passing tests: {}", passes);
-    if passes == rules.len() {
-        std::process::exit(0);
-    } else {
-        std::process::exit(1);
-    }
-}
+//     println!("Rules passing tests: {}", passes);
+//     if passes == rules.len() {
+//         std::process::exit(0);
+//     } else {
+//         std::process::exit(1);
+//     }
+// }
+
+fn main() {}
