+++ nlprule/src/compile/utils.rs
@@ -1,55 +1,3 @@
-use crate::{rules::RulesLangOptions, tokenizer::TokenizerLangOptions};
-use crate::{tokenizer::tag::TaggerLangOptions, types::*};
-use lazy_static::lazy_static;
-
-lazy_static! {
-    static ref TOKENIZER_LANG_OPTIONS: DefaultHashMap<String, TokenizerLangOptions> = {
-        serde_json::from_slice(include_bytes!(concat!(
-            env!("OUT_DIR"),
-            "/",
-            "tokenizer_configs.json"
-        )))
-        .expect("tokenizer configs must be valid JSON")
-    };
-}
-
-lazy_static! {
-    static ref RULES_LANG_OPTIONS: DefaultHashMap<String, RulesLangOptions> = {
-        serde_json::from_slice(include_bytes!(concat!(
-            env!("OUT_DIR"),
-            "/",
-            "rules_configs.json"
-        )))
-        .expect("rules configs must be valid JSON")
-    };
-}
-
-lazy_static! {
-    static ref TAGGER_LANG_OPTIONS: DefaultHashMap<String, TaggerLangOptions> = {
-        serde_json::from_slice(include_bytes!(concat!(
-            env!("OUT_DIR"),
-            "/",
-            "tagger_configs.json"
-        )))
-        .expect("tagger configs must be valid JSON")
-    };
-}
-
-/// Gets the tokenizer language options for the language code
-pub(crate) fn tokenizer_lang_options(lang_code: &str) -> Option<TokenizerLangOptions> {
-    TOKENIZER_LANG_OPTIONS.get(lang_code).cloned()
-}
-
-/// Gets the rules language options for the language code
-pub(crate) fn rules_lang_options(lang_code: &str) -> Option<RulesLangOptions> {
-    RULES_LANG_OPTIONS.get(lang_code).cloned()
-}
-
-/// Gets the tagger language options for the language code
-pub(crate) fn tagger_lang_options(lang_code: &str) -> Option<TaggerLangOptions> {
-    TAGGER_LANG_OPTIONS.get(lang_code).cloned()
-}
-
 pub(crate) use regex::from_java_regex;
 
 mod regex {
