+++ nlprule/src/components/rules/compile/structure/impls.rs
@@ -1,0 +1,375 @@
+use std::{
+    collections::hash_map::DefaultHasher,
+    hash::{Hash, Hasher},
+};
+
+use crate::utils::parallelism::MaybeParallelIterator;
+use crate::{
+    compile::{BuildInfo, Error},
+    rule::engine::{composition::*, Engine},
+    utils::regex::Regex,
+};
+use crate::{rule::disambiguation::PosFilter, types::*};
+
+impl TextMatcher {
+    pub fn new(matcher: Matcher, info: &mut BuildInfo) -> Result<Self, Error> {
+        // can not cache a matcher that depends on the graph
+        let set = if matcher.graph_id().is_some() {
+            None
+        } else if let either::Right(regex) = &matcher.matcher {
+            let mut hasher = DefaultHasher::default();
+            regex.hash(&mut hasher);
+            matcher.negate.hash(&mut hasher);
+            matcher.empty_always_false.hash(&mut hasher);
+            let matcher_hash = hasher.finish();
+
+            if let Some(set) = info.mut_regex_cache().get(&matcher_hash) {
+                set.clone()
+            } else {
+                let data: Vec<_> = info.tagger().word_store().iter().collect();
+
+                let set: DefaultHashSet<_> = data
+                    .into_maybe_par_iter()
+                    .filter_map(|(word, id)| {
+                        if matcher.is_match(word.as_str(), None, None) {
+                            Some(*id)
+                        } else {
+                            None
+                        }
+                    })
+                    .collect();
+
+                // there are some regexes which match lots of strings
+                // this cutoff is pretty arbitrary but without any threshold the size of some sets blows up
+                // the vast majority of regexes matches less than 100 strings from manual inspection
+                let set = if set.len() > 100 { None } else { Some(set) };
+                info.mut_regex_cache().insert(matcher_hash, set.clone());
+                set
+            }
+        } else {
+            None
+        };
+
+        Ok(TextMatcher { matcher, set })
+    }
+}
+
+impl PosMatcher {
+    pub fn new(matcher: Matcher, info: &mut BuildInfo) -> Result<Self, Error> {
+        let mut mask = vec![false; info.tagger().tag_store().len()];
+
+        for (word, id) in info.tagger().tag_store().iter() {
+            mask[id.value() as usize] = matcher.is_match(word.as_str(), None, None);
+        }
+
+        Ok(PosMatcher { mask })
+    }
+}
+
+impl PosFilter {
+    pub fn new(matcher: PosMatcher) -> Self {
+        PosFilter { matcher }
+    }
+}
+
+impl Regex {
+    pub fn from_java_regex(
+        java_regex_str: &str,
+        full_match: bool,
+        case_sensitive: bool,
+    ) -> Result<Self, Error> {
+        let regex_string =
+            crate::compile::utils::from_java_regex(java_regex_str, case_sensitive, full_match)?;
+
+        let regex = Regex::new(regex_string);
+        if let Err(error) = regex.try_compile() {
+            return Err(Error::Regex(error));
+        }
+
+        Ok(regex)
+    }
+}
+
+impl Engine {
+    pub fn to_graph_id(&self, id: usize) -> Result<GraphId, Error> {
+        let mut id = GraphId(id);
+
+        let map = match &self {
+            Engine::Token(engine) => &engine.composition.id_to_idx,
+            Engine::Text(_, id_to_idx) => &id_to_idx,
+        };
+
+        let max_id = *map
+            .keys()
+            .max()
+            .ok_or_else(|| Error::Unexpected("graph is empty".into()))?;
+
+        // ideally this should throw an error but LT is more lenient than nlprule
+        if !map.contains_key(&id) {
+            id = max_id;
+        }
+
+        Ok(id)
+    }
+}
+
+mod composition {
+    use super::*;
+    use crate::{
+        rule::engine::composition::{
+            AndAtom, Atom, Composition, FalseAtom, GraphId, NotAtom, OffsetAtom, OrAtom, Part,
+            Quantifier, TrueAtom,
+        },
+        utils::regex::Regex,
+    };
+
+    impl Atom {
+        fn iter_mut<'a>(&'a mut self) -> Box<dyn Iterator<Item = &'a mut Atom> + 'a> {
+            match self {
+                Atom::ChunkAtom(_)
+                | Atom::SpaceBeforeAtom(_)
+                | Atom::TextAtom(_)
+                | Atom::WordDataAtom(_)
+                | Atom::FalseAtom(_)
+                | Atom::TrueAtom(_) => Box::new(std::iter::once(self)),
+                Atom::AndAtom(x) => Box::new(x.atoms.iter_mut()),
+                Atom::OrAtom(x) => Box::new(x.atoms.iter_mut()),
+                Atom::NotAtom(x) => x.atom.iter_mut(),
+                Atom::OffsetAtom(x) => x.atom.iter_mut(),
+            }
+        }
+
+        pub fn mut_graph_ids(&mut self) -> Vec<&mut GraphId> {
+            let mut ids = Vec::new();
+
+            for atom in self.iter_mut() {
+                let id = match atom {
+                    Atom::ChunkAtom(atom) => atom.matcher.mut_graph_id(),
+                    Atom::TextAtom(atom) => atom.matcher.matcher.mut_graph_id(),
+                    Atom::WordDataAtom(atom) => atom
+                        .matcher
+                        .inflect_matcher
+                        .as_mut()
+                        .and_then(|x| x.matcher.mut_graph_id()),
+                    _ => {
+                        continue;
+                    }
+                };
+
+                if let Some(id) = id {
+                    ids.push(id);
+                }
+            }
+
+            ids
+        }
+    }
+
+    impl Matcher {
+        pub fn new_regex(regex: Regex, negate: bool, empty_always_false: bool) -> Self {
+            Matcher {
+                matcher: either::Right(regex),
+                negate,
+                case_sensitive: true, // handled by regex, should maybe be an option
+                empty_always_false,
+            }
+        }
+
+        pub fn new_string(
+            string_or_idx: either::Either<String, GraphId>,
+            negate: bool,
+            case_sensitive: bool,
+            empty_always_false: bool,
+        ) -> Self {
+            Matcher {
+                matcher: either::Left(string_or_idx),
+                negate,
+                case_sensitive,
+                empty_always_false,
+            }
+        }
+
+        pub fn graph_id(&self) -> Option<GraphId> {
+            if let either::Left(either::Right(id)) = &self.matcher {
+                Some(*id)
+            } else {
+                None
+            }
+        }
+
+        pub fn mut_graph_id(&mut self) -> Option<&mut GraphId> {
+            if let either::Left(either::Right(id)) = &mut self.matcher {
+                Some(id)
+            } else {
+                None
+            }
+        }
+    }
+
+    impl Quantifier {
+        pub fn new(min: usize, max: usize) -> Self {
+            assert!(max >= min);
+            Quantifier { min, max }
+        }
+    }
+
+    impl AndAtom {
+        pub fn and(atoms: Vec<Atom>) -> Atom {
+            let mut atoms: Vec<_> = atoms
+                .into_iter()
+                .filter(|x| !matches!(x, Atom::TrueAtom { .. }))
+                .collect();
+
+            if atoms.is_empty() {
+                (TrueAtom {}).into()
+            } else if atoms.len() == 1 {
+                atoms.remove(0)
+            } else {
+                (AndAtom { atoms }).into()
+            }
+        }
+    }
+
+    impl OrAtom {
+        pub fn or(atoms: Vec<Atom>) -> Atom {
+            let mut atoms: Vec<_> = atoms
+                .into_iter()
+                .filter(|x| !matches!(x, Atom::FalseAtom { .. }))
+                .collect();
+
+            if atoms.is_empty() {
+                (FalseAtom {}).into()
+            } else if atoms.len() == 1 {
+                atoms.remove(0)
+            } else {
+                (OrAtom { atoms }).into()
+            }
+        }
+    }
+
+    impl NotAtom {
+        pub fn not(atom: Atom) -> Atom {
+            match atom {
+                Atom::TrueAtom { .. } => FalseAtom::default().into(),
+                Atom::FalseAtom { .. } => TrueAtom::default().into(),
+                x => (NotAtom { atom: Box::new(x) }).into(),
+            }
+        }
+    }
+
+    impl OffsetAtom {
+        pub fn new(atom: Atom, offset: isize) -> Self {
+            OffsetAtom {
+                atom: Box::new(atom),
+                offset,
+            }
+        }
+    }
+
+    impl Composition {
+        pub fn new(mut parts: Vec<Part>) -> Result<Self, Error> {
+            let mut id_to_idx = DefaultHashMap::default();
+            id_to_idx.insert(GraphId(0), 0);
+            let mut current_id = 1;
+
+            for (i, part) in parts.iter().enumerate() {
+                if part.visible {
+                    id_to_idx.insert(GraphId(current_id), i + 1);
+                    current_id += 1;
+                }
+            }
+
+            let can_stop_mask = (0..parts.len())
+                .map(|i| parts[i..].iter().all(|x| x.quantifier.min == 0))
+                .collect();
+
+            for (i, part) in parts.iter_mut().enumerate() {
+                for id in part.atom.mut_graph_ids() {
+                    loop {
+                        let index = *id_to_idx.get(&id).ok_or_else(|| {
+                            Error::Unexpected(format!("id must exist in graph: {:?}", id))
+                        })?;
+
+                        // ideally this should throw an error but LT is more lenient than nlprule
+                        if index > i {
+                            *id = GraphId(id.0 - 1);
+                        } else {
+                            break;
+                        }
+                    }
+                }
+            }
+
+            Ok(Composition {
+                parts,
+                id_to_idx,
+                can_stop_mask,
+            })
+        }
+    }
+}
+
+pub mod filters {
+    use super::Error;
+    use std::collections::HashMap;
+
+    use crate::{filter::*, rule::engine::Engine, utils::regex::Regex};
+
+    trait FromArgs: Sized {
+        fn from_args(args: HashMap<String, String>, engine: &Engine) -> Result<Self, Error>;
+    }
+
+    impl FromArgs for NoDisambiguationEnglishPartialPosTagFilter {
+        fn from_args(args: HashMap<String, String>, engine: &Engine) -> Result<Self, Error> {
+            if args.contains_key("negate_postag") {
+                panic!("negate_postag not supported in NoDisambiguationEnglishPartialPosTagFilter");
+            }
+
+            Ok(NoDisambiguationEnglishPartialPosTagFilter {
+                id: engine.to_graph_id(args
+                    .get("no")
+                    .ok_or_else(|| {
+                        Error::Unexpected(
+                            "NoDisambiguationEnglishPartialPosTagFilter must have `no` argument"
+                                .into(),
+                        )
+                    })?
+                    .parse::<usize>()?)?,
+                regexp: Regex::from_java_regex(
+                    &args.get("regexp").ok_or_else(|| {
+                        Error::Unexpected(
+                        "NoDisambiguationEnglishPartialPosTagFilter must have `regexp` argument"
+                            .into(),
+                    )
+                    })?,
+                    true,
+                    true,
+                )?,
+                postag_regexp: Regex::from_java_regex(
+                    &args.get("postag_regexp").ok_or_else(|| {
+                        Error::Unexpected(
+                        "NoDisambiguationEnglishPartialPosTagFilter must have `postag_regexp` argument"
+                            .into(),
+                    )
+                    })?,
+                    true,
+                    true,
+                )?,
+                negate_postag: args.get("negate_postag").map_or(false, |x| x == "yes"),
+            })
+        }
+    }
+
+    pub fn get_filter(
+        name: &str,
+        args: HashMap<String, String>,
+        engine: &Engine,
+    ) -> Result<Filter, Error> {
+        match name {
+            "NoDisambiguationEnglishPartialPosTagFilter" => {
+                Ok(NoDisambiguationEnglishPartialPosTagFilter::from_args(args, engine)?.into())
+            }
+            _ => Err(Error::Unexpected(format!("unsupported filter {}", name))),
+        }
+    }
+}
