+++ nlprule/src/components/rules/compile/structure/mod.rs
@@ -1,7 +1,7 @@
-use fs_err::File;
 use serde::Deserialize;
-use std::io::BufReader;
-use xml::reader::EventReader;
+
+pub mod impls;
+pub mod parse;
 
 mod preprocess {
     use std::{borrow::Cow, str::FromStr};
@@ -638,155 +638,4 @@ pub enum DisambiguationRuleContainer {
     Rule(DisambiguationRule),
     RuleGroup(DisambiguationRuleGroup),
     Unification(Unification),
-}
-
-macro_rules! flatten_group {
-    ($rulegroup:expr, $category:expr) => {{
-        let group_antipatterns = if let Some(antipatterns) = $rulegroup.antipatterns {
-            antipatterns
-        } else {
-            Vec::new()
-        };
-
-        let group = Group {
-            id: $rulegroup.id,
-            default: $rulegroup.default,
-            name: $rulegroup.name,
-            n: 0,
-        };
-
-        $rulegroup
-            .rules
-            .into_iter()
-            .enumerate()
-            .map(|(i, mut rule)| {
-                if let Some(antipatterns) = &mut rule.antipatterns {
-                    antipatterns.extend(group_antipatterns.clone());
-                } else {
-                    rule.antipatterns = Some(group_antipatterns.clone());
-                }
-
-                let mut group = group.clone();
-                group.n = i;
-                (rule, Some(group), $category.clone())
-            })
-            .collect::<Vec<_>>()
-    }};
-}
-
-type GrammarRuleReading = (Rule, Option<Group>, Option<Category>);
-type DisambiguationRuleReading = (DisambiguationRule, Option<Group>, Option<Category>);
-
-pub fn read_rules<P: AsRef<std::path::Path>>(
-    path: P,
-) -> Vec<Result<GrammarRuleReading, serde_xml_rs::Error>> {
-    let file = File::open(path.as_ref()).unwrap();
-    let file = BufReader::new(file);
-
-    let sanitized = preprocess::sanitize(file, &["suggestion"]);
-    let rules = preprocess::extract_rules(sanitized.as_bytes());
-
-    let mut unifications = Vec::new();
-
-    let rules: Vec<_> = rules
-        .into_iter()
-        .map(|(xml, category)| {
-            let mut out = Vec::new();
-
-            let deseralized = RuleContainer::deserialize(&mut serde_xml_rs::Deserializer::new(
-                EventReader::new(xml.as_bytes()),
-            ));
-
-            out.extend(match deseralized {
-                Ok(rule_container) => match rule_container {
-                    RuleContainer::Rule(rule) => {
-                        vec![Ok((rule, None, category))]
-                    }
-                    RuleContainer::RuleGroup(rule_group) => flatten_group!(rule_group, category)
-                        .into_iter()
-                        .map(Ok)
-                        .collect(),
-                    RuleContainer::Unification(unification) => {
-                        unifications.push(unification);
-
-                        vec![]
-                    }
-                },
-                Err(err) => vec![Err(err)],
-            });
-            out
-        })
-        .flatten()
-        .collect();
-
-    rules
-        .into_iter()
-        .map(|result| match result {
-            Ok(mut x) => {
-                x.0.unifications = Some(unifications.clone());
-
-                Ok(x)
-            }
-            Err(x) => Err(x),
-        })
-        .collect()
-}
-
-pub fn read_disambiguation_rules<P: AsRef<std::path::Path>>(
-    path: P,
-) -> Vec<Result<DisambiguationRuleReading, serde_xml_rs::Error>> {
-    let file = File::open(path.as_ref()).unwrap();
-    let file = BufReader::new(file);
-
-    let sanitized = preprocess::sanitize(file, &[]);
-    let rules = preprocess::extract_rules(sanitized.as_bytes());
-
-    let mut unifications = Vec::new();
-
-    let rules: Vec<_> = rules
-        .into_iter()
-        .map(|(xml, _)| {
-            let mut out = Vec::new();
-
-            let deseralized = DisambiguationRuleContainer::deserialize(
-                &mut serde_xml_rs::Deserializer::new(EventReader::new(xml.as_bytes())),
-            );
-
-            let category: Option<Category> = None;
-
-            out.extend(match deseralized {
-                Ok(rule_container) => match rule_container {
-                    DisambiguationRuleContainer::Rule(rule) => {
-                        vec![Ok((rule, None, category))]
-                    }
-                    DisambiguationRuleContainer::RuleGroup(rule_group) => {
-                        flatten_group!(rule_group, category)
-                            .into_iter()
-                            .map(Ok)
-                            .collect()
-                    }
-                    DisambiguationRuleContainer::Unification(unification) => {
-                        unifications.push(unification);
-
-                        vec![]
-                    }
-                },
-                Err(err) => vec![Err(err)],
-            });
-            out
-        })
-        .flatten()
-        .collect();
-
-    rules
-        .into_iter()
-        .map(|result| match result {
-            Ok(mut x) => {
-                x.0.unifications = Some(unifications.clone());
-
-                Ok(x)
-            }
-            Err(x) => Err(x),
-        })
-        .collect()
 }
